+++ packages/core/src/node/env-variables/env-variables-server.ts
@@ -14,21 +14,18 @@
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  ********************************************************************************/
 
-import * as os from 'os';
+import { join } from 'path';
+import { homedir } from 'os';
 import { injectable } from 'inversify';
 import { EnvVariable, EnvVariablesServer } from '../../common/env-variables';
 import { isWindows } from '../../common/os';
 import { FileUri } from '../file-uri';
 
-const THEIA_DATA_FOLDER = '.theia';
-
-const WINDOWS_APP_DATA_DIR = 'AppData';
-const WINDOWS_ROAMING_DIR = 'Roaming';
-
 @injectable()
 export class EnvVariablesServerImpl implements EnvVariablesServer {
 
     protected readonly envs: { [key: string]: EnvVariable } = {};
+    protected readonly configDirUri = FileUri.create(join(homedir(), '.theia')).toString();
 
     constructor() {
         const prEnv = process.env;
@@ -52,26 +49,8 @@ export class EnvVariablesServerImpl implements EnvVariablesServer {
         return this.envs[key];
     }
 
-    async getUserHomeFolder(): Promise<string> {
-        return FileUri.create(os.homedir()).toString();
-    }
-
-    async getDataFolderName(): Promise<string> {
-        return THEIA_DATA_FOLDER;
-    }
-
-    async getUserDataFolder(): Promise<string> {
-        return FileUri.create(await this.getUserHomeFolder()).resolve(await this.getDataFolderName()).toString();
-    }
-
-    async getAppDataFolder(): Promise<string> {
-        const dataFolderUriBuilder = FileUri.create(await this.getUserHomeFolder());
-        if (isWindows) {
-            dataFolderUriBuilder.resolve(WINDOWS_APP_DATA_DIR);
-            dataFolderUriBuilder.resolve(WINDOWS_ROAMING_DIR);
-        }
-        dataFolderUriBuilder.resolve(await this.getDataFolderName());
-        return dataFolderUriBuilder.toString();
+    async getConfigDirUri(): Promise<string> {
+        return this.configDirUri;
     }
 
 }