+++ packages/plugin-ext/src/common/plugin-api-rpc.ts
@@ -79,7 +79,6 @@ import { ArgumentProcessor } from '../plugin/command-registry';
 import { MaybePromise } from '@theia/core/lib/common/types';
 import { QuickOpenItem, QuickOpenItemOptions } from '@theia/core/lib/common/quick-open-model';
 import { QuickTitleButton } from '@theia/core/lib/common/quick-open-model';
-import { EnvVariable } from '@theia/core/lib/common/env-variables';
 
 export interface PreferenceData {
     [scope: number]: any;
@@ -993,13 +992,7 @@ export interface DocumentsMain {
 
 export interface EnvMain {
     $getEnvVariable(envVarName: string): Promise<string | undefined>;
-    $getAllEnvVariables(): Promise<EnvVariable[]>
     $getClientOperatingSystem(): Promise<theia.OperatingSystem>;
-    $getExecPath(): Promise<string>
-    $getUserHomeFolderPath(): Promise<string>
-    $getDataFolderName(): Promise<string>
-    $getUserDataFolderPath(): Promise<string>
-    $getAppDataPath(): Promise<string>
 }
 
 export interface PreferenceRegistryMain {